MAJOR CHANGES The material management system has been completely overhauled, now potentially supporting all material maps used by raylib.
It is now possible to use all values defined directly in the model materials or to use and modify internal default materials. This is useful for the upcoming raylib release, which allows defining normals for primitives, enabling them to be illuminated.

The `RLG_MaterialMap` enum has been removed in favor of using raylib's `MaterialMapIndex` enum directly, as all materials can now be supported.

An internal struct has been added for managing locations, primarily serving the lighting shader.

The lighting shader has been partially rewritten to use a struct for materials, making it easier to understand and slightly reducing code complexity.

Ambient occlusion calculation has been improved. The value obtained from the AO map is directly applied to the ambient color, and the value associated with the material `MATERIAL_MAP_OCCLUSION` serves as a scaling factor for light influence. We interpolate a value between 1.0 and the value obtained from the AO map using the material's associated value to scale the occlusion applied to light accumulation.

Many function names and their usage have been revised to be simpler and more intuitive.

And probably other minor changes made alongside all this.